refactor: 정렬 수정
