+++ src/components/Common/Dropdown/Dropdown.jsx
@@ -2,17 +2,12 @@ import styles from "./Dropdown.module.css";
 import arrowDown from "../../../assets/ic_toggle.svg";
 import { useState, useRef, useEffect } from "react";
 
-const options = {
-  "View My Startup 누적 투자 금액 높은 순": ["sim_invest", "desc"],
-  "View My Startup 누적 투자 금액 낮은 순": ["sim_invest", "asc"],
-  "실제 누적 투자 금액 높은 순": ["actual_invest", "desc"],
-  "실제 누적 투자 금액 낮은 순": ["actual_invest", "asc"],
-};
-
-export default function Dropdown({ setSortOrder }) {
+export default function Dropdown({ sortOptions, setSortOrder, order, sort }) {
   const [isOpen, setIsOpen] = useState(false);
-  const [currentLabel, setCurrentLabel] = useState(Object.keys(options)[0]);
   const dropdownRef = useRef(null);
+  const currentLabel = Object.entries(sortOptions).find(
+    ([, value]) => value[0] === order && value[1] === sort
+  )?.[0];
 
   // 메뉴 외부 클릭 감지
   useEffect(() => {
@@ -32,23 +27,20 @@ export default function Dropdown({ setSortOrder }) {
   }, [dropdownRef]);
 
   const handleOptionClick = (val) => {
-    const [orderValue, sortValue] = options[val];
+    const [orderValue, sortValue] = sortOptions[val];
     setSortOrder(orderValue, sortValue);
-    setCurrentLabel(val);
     setIsOpen(false);
   };
 
   return (
     <div className={styles.menu} ref={dropdownRef}>
       <div className={styles.selected} onClick={() => setIsOpen(!isOpen)}>
-        <p className={styles.currentLabel}>
-          {currentLabel || Object.keys(options)[0]}
-        </p>
+        <p className={styles.currentLabel}>{currentLabel}</p>
         <img className={styles.icon} src={arrowDown} alt="드롭다운 아이콘" />
       </div>
       {isOpen && (
         <ul className={styles.options}>
-          {Object.keys(options).map((option, idx) => (
+          {Object.keys(sortOptions).map((option, idx) => (
             <li
               key={idx}
               className={styles.option}
